refactor: move wizard components to $lib/wizard
