+++ src/lib/wizard/WizardNavItem.svelte
